Fix compiler warning: implicit conversion loses integer precision
grisbi/src/utils_str.c:1083:13: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
    size1 = g_utf8_strlen (basename, -1);
          ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~
grisbi/src/utils_str.c:1085:13: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
    size2 = g_utf8_strlen (dirname, -1);
          ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~
grisbi/src/utils_str.c:1106:17: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        size3 = g_utf8_strlen (tmp_dir, -1);
              ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~
grisbi/src/utils_str.c:1155:21: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
            size1 = g_utf8_strlen (basename, -1);
                  ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~
grisbi/src/utils_str.c:1164:17: warning: implicit conversion loses integer precision: 'glong' (aka 'long') to 'gint' (aka 'int') [-Wshorten-64-to-32]
        size3 = g_utf8_strlen (tmp_base, -1);
              ~ ^~~~~~~~~~~~~~~~~~~~~~~~~~~~
grisbi/src/utils_str.c:1150:19: warning: implicit conversion loses integer precision: 'unsigned long' to 'gint' (aka 'int') [-Wshorten-64-to-32]
        n = size1 / trunc;
          ~ ~~~~~~^~~~~~~

+++ src/utils_str.c
@@ -1071,10 +1071,10 @@ gchar *utils_str_break_filename (const gchar *string,
     gchar *ptr = NULL;
     gchar *separator;
     gint i = 1;
-    gint n = 0;
-    gint size1;
-    gint size2;
-    gint size3;
+    ssize_t n = 0;
+    ssize_t size1;
+    ssize_t size2;
+    ssize_t size3;
 
     if ((gint) g_utf8_strlen (string, -1) <= trunc)
         return g_strdup (string);
